Add Images, Add images to HTML file, Adjust CSS
I added the images folder for use in the HTML file. I altered the image
tag to use the images files. Alted the CSS to change the secondary
button styling and alter the radius on the primary button. Alterd the
text in the call-to-action to be white so the contrast is better
against the background color.